# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  controllers.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  controllers.TokenRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/login:
    post:
      description: Generate authentication token
      parameters:
      - description: Token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Token generation
      tags:
      - Auth
  /auth/refresh-token:
    post:
      description: Generate authentication token
      parameters:
      - description: Refrest Token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Token generation
      tags:
      - Auth
  /auth/register:
    post:
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Register user
      tags:
      - Auth
  /ping:
    get:
      description: Ping system to check if it works
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Ping
      tags:
      - Diagnostic
  /workout/:userId/generate:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Generate workout plan
      tags:
      - Workout
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"